% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{m2r}
\alias{m2r}
\title{Transform median time into rate}
\usage{
m2r(x)
}
\arguments{
\item{x}{The median time in months to be transformed into rate}
}
\description{
This helper function calculates the hazard rate per day of an exponential 
distribution from the median given in months.
}
