% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{nph_gui}
\alias{nph_gui}
\title{Launch a GUI (shiny app) for the nph package}
\usage{
nph_gui()
}
\description{
Launch a GUI (shiny app) for the nph package
}
\details{
The packages \code{shinycssloaders}, \code{formatR} and \code{styler} are required for correct display of the GUI.
The package \code{rmarkdown} with access to pandoc is required to save reports.
}
