% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultane_Inferenz_Funktionen_nph_1-3.R
\docType{data}
\name{pembro}
\alias{pembro}
\title{Reconstructed Data Set Based On Survival Curves In Burtess et al. 2019}
\format{
A data frame.
}
\usage{
data(pembro)
}
\description{
The data set was approximately reconstructed from the survival curves shown in Figure 2D of Burtness et al. 2019 (see references). It contains survival times and event #' indicator under treatment with pembrolizumab (group 1) versus cetuximab with chemotherapy (group 0).
}
\examples{
data(pembro)
head(pembro)
}
\references{
Burtness, B., Harrington, K. J., Greil, R., Soulières, D., Tahara, M., de Castro Jr, G., ... & Abel, E. (2019). Pembrolizumab alone or with chemotherapy
versus cetuximab with chemotherapy for recurrent or metastatic squamous cell carcinoma of the head and neck (KEYNOTE-048): a randomised, open-label,
phase 3 study. The Lancet, 394(10212), 1915-1928.
}
\keyword{datasets}
