% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions_1-1.R
\name{plot_shhr}
\alias{plot_shhr}
\title{Plot of survival, hazard and hazard ratio of two groups as a function of time}
\usage{
plot_shhr(A, B, main = "", xmax = NULL, ymax_haz = NULL, ymax_hr = NULL)
}
\arguments{
\item{A}{An object of class \code{mixpch}, resembling the survival function in treatment group 0}

\item{B}{An object of class \code{mixpch}, resembling the survival function in treatment group 1}

\item{main}{An overall title for the plot}

\item{xmax}{A maximum value for the x-axis. The plot is drawn using xlim = c(0, xmax)}

\item{ymax_haz}{A maximum value for the y-axis for the hazards plot. The plot is drawn using ylim = c(0, ymax_haz)}

\item{ymax_hr}{A maximum value for the y-axis for the hazards ratio plot. The plot is drawn using ylim = c(0, ymax_hr)}
}
\description{
A convenience function that uses the generic plot function in the nph package to
plot the three functions in a layout of 3 columns and 1 row.
}
