% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultane_Inferenz_Funktionen_nph_1-3.R
\name{print.nphparams}
\alias{print.nphparams}
\title{Print nphparams Objects}
\usage{
\method{print}{nphparams}(x, ...)
}
\arguments{
\item{x}{an object of class \code{nphparams}.}

\item{...}{further arguments, not used.}
}
\description{
Prints the results table of an nphparams object.
}
\details{
Estiamtes corresponding to differences at a log scale are transformed by taking exp(), and are labelled as ratios. I.e. differnces in log urvival probabilites,
differences in log quantiles, cloglog survival differences (equivalent to the log cumulative hazard ratio), log average hazard ratios or Cox model log hazard ratioss are
transformed to survival probability ratios, quantile ratios, cumulative hazard ratios, average hazard ratios or Cox model hazard ratios, respectively. In the output,
the standard error at the backtransformed scale is calculated by the delta-method. Confidence interval bounds are calculated at the log-scale, though, and then
transformed by taking exp().
}
\examples{
data(pembro)
set1<-nphparams(time=time, event=event, group=group,data=pembro,
param_type=c("score","S"),
param_par=c(3.5,2),
param_alternative=c("less","greater"),
closed_test=TRUE,alternative_test="one.sided")
print(set1)
plot(set1,trt_name="Pembrolizumab",ctr_name="Cetuximab")

set2<-nphparams(time=time, event=event, group=group, data=pembro,  
param_type=c("S","S","S","Q","RMST"),
param_par=c(0.5,1,2,0.5,3.5))
print(set2)
plot(set2,showlines=TRUE,show_rmst_diff=TRUE)

}
\seealso{
\code{\link{nphparams}}, \code{\link{plot.nphparams}}
}
\author{
Robin Ristl, \email{robin.ristl@meduniwien.ac.at}
}
